Introduce base class for hitobjects that can be strong
