+++ osu.Game.Rulesets.Taiko.Tests/TaikoBeatmapConversionTest.cs
@@ -35,7 +35,7 @@ namespace osu.Game.Rulesets.Taiko.Tests
                 IsCentre = (hitObject as Hit)?.Type == HitType.Centre,
                 IsDrumRoll = hitObject is DrumRoll,
                 IsSwell = hitObject is Swell,
-                IsStrong = ((TaikoHitObject)hitObject).IsStrong
+                IsStrong = (hitObject as TaikoStrongHitObject)?.IsStrong == true
             };
         }
