+++ osu.Game.Rulesets.Taiko/Objects/TaikoHitObject.cs
@@ -1,9 +1,6 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using System;
-using System.Threading;
-using osu.Framework.Bindables;
 using osu.Game.Rulesets.Judgements;
 using osu.Game.Rulesets.Objects;
 using osu.Game.Rulesets.Scoring;
@@ -18,47 +15,6 @@ namespace osu.Game.Rulesets.Taiko.Objects
         /// Default size of a drawable taiko hit object.
         /// </summary>
         public const float DEFAULT_SIZE = 0.45f;
-
-        /// <summary>
-        /// Scale multiplier for a strong drawable taiko hit object.
-        /// </summary>
-        public const float STRONG_SCALE = 1.4f;
-
-        /// <summary>
-        /// Default size of a strong drawable taiko hit object.
-        /// </summary>
-        public const float DEFAULT_STRONG_SIZE = DEFAULT_SIZE * STRONG_SCALE;
-
-        public readonly Bindable<bool> IsStrongBindable = new BindableBool();
-
-        /// <summary>
-        /// Whether this <see cref="TaikoHitObject"/> can be made a "strong" (large) hit.
-        /// </summary>
-        public virtual bool CanBeStrong => true;
-
-        /// <summary>
-        /// Whether this HitObject is a "strong" type.
-        /// Strong hit objects give more points for hitting the hit object with both keys.
-        /// </summary>
-        public bool IsStrong
-        {
-            get => IsStrongBindable.Value;
-            set
-            {
-                if (value && !CanBeStrong)
-                    throw new InvalidOperationException($"Object of type {GetType()} cannot be strong");
-
-                IsStrongBindable.Value = value;
-            }
-        }
-
-        protected override void CreateNestedHitObjects(CancellationToken cancellationToken)
-        {
-            base.CreateNestedHitObjects(cancellationToken);
-
-            if (IsStrong)
-                AddNested(new StrongNestedHitObject { StartTime = this.GetEndTime() });
-        }
 
         public override Judgement CreateJudgement() => new TaikoJudgement();
